pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kb-builder"
version = "0.1.0"
description = "Game design knowledge base builder: converts planning docs (xlsx/docx) into a structured wiki with entity cross-references, concept indices, and an interactive knowledge graph."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "markitdown[docx,xlsx]>=0.1.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["scripts"]
exclude = ["old", "knowledge", "kb_builder", "wiki_specs", "llm-wiki-agent*", "references", "references*"]

[project.optional-dependencies]
graph = [
    "networkx>=3.0",
]
llm = [
    "litellm>=1.0",
    "openai>=1.0",
]
dev = [
    "kb-builder[graph,llm]",
]
